Merge branch 'evcore-common-master'

cordova-sqlite-evcore-common-free compile-time option fixes for iOS/macOS:
- SQLITE_THREADSAFE=2 on iOS/macOS to avoid possible malformed database due
  to multithreaded access ref: litehelpers/Cordova-sqlite-storage#703
- Suppress warnings when building sqlite3.c (iOS/macOS)
- Remove unwanted SQLITE_OMIT_BUILTIN_TEST option (iOS/macOS)

Additional doc fixes

+++ spec/www/spec/db-open-close-delete-test.js
@@ -1130,8 +1130,7 @@ var mytests = function() {
           }
         }
 
-        test_it(suiteName + ' test sqlitePlugin.deleteDatabase()', function () {
-          stop();
+        it(suiteName + ' test sqlitePlugin.deleteDatabase()', function (done) {
           var db = openDatabase("DB-Deletable", "1.0", "Demo", DEFAULT_SIZE);
 
           function createAndInsertStuff() {
@@ -1142,9 +1141,14 @@ var mytests = function() {
                 tx.executeSql('INSERT INTO test VALUES (?)', ['foo']);
               });
             }, function (err) {
-              ok(false, 'create and insert tx failed with ERROR: ' + JSON.stringify(err));
+              // NOT EXPECTED:
               console.log('create and insert tx failed with ERROR: ' + JSON.stringify(err));
-              start();
+              expect(false).toBe(true);
+              expect(err).toBeDefined();
+              expect(err.message).toBeDefined();
+              expect(err.message).toBe('--');
+              done();
+
             }, function () {
               // check that we can read it
               db.transaction(function(tx) {
@@ -1152,9 +1156,13 @@ var mytests = function() {
                   equal(res.rows.item(0).name, 'foo');
                 });
               }, function (err) {
-                ok(false, 'SELECT tx failed with ERROR: ' + JSON.stringify(err));
+                // NOT EXPECTED:
                 console.log('SELECT tx failed with ERROR: ' + JSON.stringify(err));
-                start();
+                expect(false).toBe(true);
+                expect(err).toBeDefined();
+                expect(err.message).toBeDefined();
+                expect(err.message).toBe('--');
+                done();
               }, function () {
                 deleteAndConfirmDeleted();
               });
@@ -1169,32 +1177,38 @@ var mytests = function() {
               db.transaction(function (tx) {
                 tx.executeSql('SELECT name FROM test', []);
               }, function (err) {
-                ok(true, 'got an expected transaction error');
+                // EXPECTED RESULT:
+                expect(err).toBeDefined();
+                expect(err.message).toBeDefined();
                 testDeleteError();
               }, function () {
+                // SUCCESS CALLBACK NOT EXPECTED:
                 console.log('UNEXPECTED SUCCESS: expected a transaction error');
-                ok(false, 'expected a transaction error');
-                start();
+                expect(false).toBe(true);
+                done();
               });
             }, function (err) {
+              // NOT EXPECTED - DO NOT IGNORE ON ANY PLATFORM:
               console.log("ERROR: " + JSON.stringify(err));
-              // XXX TBD IGNORE delete error on Windows:
-              if (isWindows) return start();
-              ok(false, 'error: ' + err);
-              start();
+              expect(false).toBe(true);
+              expect(err).toBeDefined();
+              expect(err.message).toBeDefined();
+              expect(err.message).toBe('--');
+              done();
             });
           }
 
           function testDeleteError() {
             // should throw an error if the db doesn't exist
             deleteDatabase("Foo-Doesnt-Exist", function () {
+              // SUCCESS CALLBACK NOT EXPECTED:
               console.log('UNEXPECTED SUCCESS: expected a delete error');
-              ok(false, 'expected error');
-              start();
+              expect(false).toBe(true);
+              done();
             }, function (err) {
-              ok(!!err, 'got error like we expected');
-
-              start();
+              // EXPECTED RESULT:
+              expect(err).toBeDefined();
+              done();
             });
           }
 
@@ -1311,7 +1325,7 @@ var mytests = function() {
 
         test_it(suiteName + ' database.close (immediately after open) calls its success callback', function () {
           // TBD POSSIBLY BROKEN on iOS/macOS due to current background processing implementation:
-          if (!isAndroid && !isWindows && !isWP8) pending('POSSIBLY BROKEN on iOS/macOS (background processing implementation)');
+          if (!isAndroid && !isWindows && !isWP8) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
 
           // asynch test coming up
           stop(1);
@@ -1471,16 +1485,46 @@ var mytests = function() {
       });
     }
 
-    describe('repeated open/close/delete test(s)', function() {
-      var scenarioName = isAndroid ? 'Plugin-implementation-default' : 'Plugin';
-      var suiteName = scenarioName + ': ';
+    for (var i=0; i<pluginScenarioCount; ++i) {
+
+      describe(pluginScenarioList[i] + ': repeated open/close/delete test(s)', function() {
+        var scenarioName = pluginScenarioList[i];
+        var suiteName = scenarioName + ': ';
+        var isImpl2 = (i === 1);
 
         // NOTE: MUST be defined in function scope, NOT outer scope:
         var openDatabase = function(first, second, third) {
+          if (first.constructor === String ) throw new Error('string not expected here');
+
+          // androidDatabaseImplementation: 2 (builtin android.database implementation):
+          if (isImpl2) {
+            var dbname = first.name;
+            return window.sqlitePlugin.openDatabase({
+              name: 'i2-'+dbname,
+              // database location setting needed here (value ignored on Android):
+              location: 'default',
+              androidDatabaseImplementation: 2,
+              androidLockWorkaround: 1
+            }, second, third);
+          }
+
           return window.sqlitePlugin.openDatabase(first, second, third);
         }
 
         var deleteDatabase = function(first, second, third) {
+          if (first.constructor === String ) throw new Error('string not expected here');
+
+          // androidDatabaseImplementation: 2 (builtin android.database implementation):
+          if (isImpl2) {
+            var dbname = first.name;
+            return window.sqlitePlugin.deleteDatabase({
+              name: 'i2-'+dbname,
+              // database location setting needed here (value ignored on Android):
+              location: 'default',
+              androidDatabaseImplementation: 2
+            }, second, third);
+          }
+
           window.sqlitePlugin.deleteDatabase(first, second, third);
         }
 
@@ -1495,23 +1539,23 @@ var mytests = function() {
           var db1 = openDatabase(dbargs, function () {
             var db2 = openDatabase(dbargs, function () {
               db1.readTransaction(function(tx1) {
-                tx1.executeSql('SELECT 1', [], function(tx1d, results) {
+                tx1.executeSql('SELECT 1', [], function(tx_ignored, results) {
                   ok(true, 'db1 transaction working');
                   start(1);
-                }, function(ignored, error) {
+                }, function(tx_ignored, error) {
                   ok(false, error);
                 });
               }, function(error) {
                 ok(false, error);
               });
               db2.readTransaction(function(tx2) {
-                tx2.executeSql('SELECT 1', [], function(tx2d, results) {
+                tx2.executeSql('SELECT 1', [], function(tx_ignored, results) {
                   ok(true, 'db2 transaction working');
                   start(1);
-                }, function(ignored, error) {
+                }, function(tx_ignored, error) {
                   ok(false, error);
                 });
-              }, function(error) {
+              }, function(tx_ignored, error) {
                 ok(false, error);
               });
             }, function (error) {
@@ -1617,8 +1661,7 @@ var mytests = function() {
         // XXX SEE BELOW: repeat scenario but wait for open callback before close/delete/reopen
         // Needed to support some large-scale applications:
         test_it(suiteName + ' immediate close, then delete then re-open allows subsequent queries to run', function () {
-          // TBD POSSIBLY BROKEN on iOS/macOS due to current background processing implementation:
-          if (!isAndroid && !isWindows && !isWP8) pending('POSSIBLY BROKEN on iOS/macOS (background processing implementation)');
+          if (!isAndroid && !isWindows && !isWP8) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
 
           var dbName = "Immediate-close-delete-Reopen.db";
           var dbargs = {name: dbName, location: 'default'};
@@ -1766,6 +1809,8 @@ var mytests = function() {
         test_it(suiteName + ' repeatedly open and close database faster (5x)', function () {
           // TBD CURRENTLY BROKEN on iOS/macOS due to current background processing implementation:
           if (!isAndroid && !isWindows && !isWP8) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          // TBD ???:
+          if (isAndroid && isImpl2) pending('FAILS on builtin android.database implementation (androidDatabaseImplementation: 2)');
 
           var dbName = 'repeatedly-open-and-close-faster-5x.db';
           var dbargs = {name: dbName, location: 'default'};
@@ -1886,8 +1931,10 @@ var mytests = function() {
 
         // Needed to support some large-scale applications:
         test_it(suiteName + ' repeatedly open and delete database faster (5x)', function () {
-          // TBD CURRENTLY BROKEN on iOS/macOS due to current background processing implementation:
-          if (!isAndroid && !isWindows && !isWP8) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          // TBD POSSIBLY BROKEN on iOS/macOS ...
+          // if (!isAndroid && !isWindows && !isWP8) pending(...);
+          // TBD CURRENTLY BROKEN DUE TO BUG 666 WORKAROUND SOLUTION
+          pending('CURRENTLY BROKEN DUE TO BUG 666 WORKAROUND SOLUTION');
 
           var dbName = 'repeatedly-open-and-delete-faster-5x.db';
           var dbargs = {name: dbName, location: 'default'};
@@ -1943,7 +1990,8 @@ var mytests = function() {
           });
         });
 
-    });
+      });
+    }
 
   });
